Create Project #6: Description
What you will see in this project 👉🏻 Creation of a beginner level report with a step-by-step explanation to create it, following the phases of Microsoft Power BI for data analysis: Get Data (GD), Data Preparation (DP) with PowerQuery or Query Editor, Data Modeling (DM) with DAX, Data Visualization (DV), Data Reporting (DR).

Project: Comparison of Total Income vs Previous Years

This project aims to carry out an exhaustive analysis of the income generated by the company, comparing the current results with those of previous years. Through interactive visualizations and detailed analysis, it seeks to identify trends, patterns and areas of improvement to optimize commercial and distribution strategies.

Areas of Analysis:
• Total Income: The analysis begins with an overview of the Total Income obtained during the current period and its comparison with the income of previous years. This key indicator allows observing the growth or decrease in income, providing a clear measure of the company's overall performance.

• Total Revenue by Product Subcategory: A detailed breakdown of revenue is made by Product Subcategories. This helps to identify which products or product groups are generating the most revenue and how they compare to previous years' results. This analysis allows the impact of product strategies to be assessed and offers to be adjusted to maximize revenue.

• Total Revenue by Region: Revenue is analyzed by Region, providing a clear view of how sales are distributed geographically. This analysis allows the performance of the company in different markets to be observed and compared to revenue from previous years. It helps to identify regions with the greatest potential or areas that require additional strategic focus.

Project Objective:
The main objective is to offer a clear and accurate view of the evolution of revenue, both globally and segmented by product subcategory and region. The results will allow the company to make decisions based on historical data and optimize its approach to maximize revenue in the future.