Dizteku Financial Dashboard
What you will see in this project 👉🏻 Creation of a beginner-level report with a step-by-step explanation of how to do it, following the phases of Microsoft Power BI for data analysis: Get Data (GD), Data Preparation (DP) with PowerQuery or Query Editor, Data Modeling (DM) with DAX, Data Visualization (DV), Data Reporting (DR).

Project: Dizteku Financial Dashboard

This project aims to develop an interactive Financial Dashboard for Dizteku, which consolidates the company's key financial metrics, allowing for a real-time analysis of economic performance. The dashboard displays total revenue, expenses, profits, profit margins, and the monthly evolution of profits, providing a useful tool for strategic decision-making and financial planning.

Areas of Analysis:
• Total Revenue: Total Revenue is one of the main financial indicators of the dashboard. It shows the volume of revenue generated during a specific period, allowing users to monitor the financial health of the company and evaluate its performance against financial targets.

• Total Expenses: A detailed breakdown of the company's Total Expenses is presented, including both fixed and variable costs. This metric is essential for assessing cost control and operational efficiency, allowing managers to identify areas where expenses could be optimized.

• Total Profit: Total Profit reflects the net profitability obtained, after deducting expenses from revenue. This indicator is essential for assessing the company's overall success in generating profits, providing a clear view of its economic performance.

• Profit Margin Percentage: The Profit Margin Percentage shows the company's profitability in percentage terms, i.e. what percentage of revenue is converted into net profit. This indicator is key for assessing Dizteku's operational efficiency and financial sustainability.

• Total Profit by Month: A monthly display of Total Profit allows you to observe profitability fluctuations and trends over time. This metric makes it easy to identify seasonal patterns, as well as the impact of external or internal factors on financial results.

Project Objective:
The main objective of this project is to offer an interactive visual tool that allows Dizteku's financial managers to track key metrics in real time, facilitating informed decision-making and resource optimization. The dashboard provides a comprehensive view of the financial situation, helping to manage the flow of income and expenses to maximize profitability.