Create Project #8: Description
What you will see in this project 👉🏻 Creation of a beginner-level report with a step-by-step explanation of how to do it, following the phases of Microsoft Power BI for data analysis: Get Data (GD), Data Preparation (DP) with PowerQuery or Query Editor, Data Modeling (DM) with DAX, Data Visualization (DV), Data Reporting (DR).

Project: Basic Twitch Report

This project aims to present a Basic Twitch Report focused on general video game metrics and the analysis of the average number of viewers by time intervals. Through this report, we seek to offer a clear view of the behavior and popularity of video games on the streaming platform, as well as identify trends in the number of viewers over time.

Areas of Analysis:
• General Video Game Metrics: A general analysis of the most popular and broadcast video games on Twitch is presented, with key metrics such as the total number of streams, the number of hours watched, and viewer interaction. This analysis allows you to see which games are most engaging for the Twitch audience, providing valuable information for developers and content creators.

• Average Viewership by Time Interval Metric: A key part of the report focuses on the average number of viewers during specific time intervals, allowing you to identify spikes in popularity and patterns in viewing. This analysis helps you understand how viewership varies at different times of the day or week, providing data that can be used to optimize streaming schedules and improve visibility for certain games.

Project Goal:
The main goal of this project is to offer a simple yet informative report on the performance of video games on Twitch, focusing on viewership metrics and viewing patterns. This information can be used by both content creators and game developers to make strategic decisions about when and how to stream their content, as well as to identify the most popular games among Twitch users.